backend/requirements.txt
# Runtime
fastapi~=0.119.1
pydantic~=2.12.3
email-validator~=2.3.0
sqlalchemy~=2.0.44
asyncpg~=0.30.0
uvicorn[standard]~=0.38.0
PyJWT~=2.9.0

# Dev (lint/type/format)
ruff~=0.14.1
mypy~=1.18.2
